Add LoginPage to Todo App (#20)
This currently stores the login result (email and user ID) in
SharedPreferences. If not present we show a LoginPage where we allow the
user to enter an email. When submitting we go to `/serve/login` to get a
user ID. We then pass this user id as the `clientViewAuth`.

Logging out jsut clear the local storage and the state and shows the
LoginPage again.

Towards rocicorp/replicache#25