Switch redo sample app and tests to use production diff server.
For the sample app, I need this in order to demo offline. It's more
difficult to simulate local hosts being unavailable than the public
internet.

I also think this is a better development experience, one less
thingy that needs to be running.

This does make more obvious that diff-server must maintain a
backward-compatible HTTP interface to repm. However, that was
always the case - there was no version pinning between repm and
diff-server. Also we will need this to be the case once we are
being used.

If it turns out to be useful for some reason to use the local
diff-server we could add additional flags to test_server and/or
redo.